(PC-3164): Add condition for user can book free offers
