+++ repository/user_queries.py
@@ -158,3 +158,7 @@ def _sanitized_string(value: str) -> Function:
     sanitized = func.unaccent(sanitized)
     sanitized = func.lower(sanitized)
     return sanitized
+
+
+def find_user_by_id(user_id: id) -> User:
+    return User.query.get(user_id)
